Rename go app (#98)
* rename go app
* image name: remove -gs suffix
* update repo name in docs and examples

---------
