+++ .circleci/config.yml
@@ -10,7 +10,7 @@ workflows:
      - architect/push-to-registries:
         context: architect
         name: push-to-registries
-        image: "giantswarm/loki-multi-tenant-proxy-gs"
+        image: "giantswarm/grafana-multi-tenant-proxy"
         dockerfile: "build/package/Dockerfile-with-tests"
         filters:
           tags:
