+++ .pre-commit-config.yaml
@@ -28,4 +28,4 @@ repos:
         # timeout is needed for CI
         args: [ -E, gosec, -E, goconst, -E, govet, --timeout, 300s ]
       - id: go-imports
-        args: [ -local, github.com/giantswarm/loki-multi-tenant-proxy ]
+        args: [ -local, github.com/giantswarm/grafana-multi-tenant-proxy ]
