+++ examples/kubernetes/loki-multi-tenant-proxy.yaml
@@ -2,10 +2,10 @@
 apiVersion: v1
 kind: Secret
 metadata:
-  name: loki-multi-tenant-proxy-auth-config
+  name: grafana-multi-tenant-proxy-auth-config
   namespace: grafana
   labels:
-    app: loki-multi-tenant-proxy
+    app: grafana-multi-tenant-proxy
 stringData:
   authn.yaml: |-
     users:
@@ -19,7 +19,7 @@ stringData:
 apiVersion: v1
 kind: Service
 metadata:
-  name: loki-multi-tenant-proxy
+  name: grafana-multi-tenant-proxy
   namespace: grafana
 spec:
   ports:
@@ -27,34 +27,34 @@ spec:
     protocol: TCP
     targetPort: 3100
   selector:
-    app: loki-multi-tenant-proxy
+    app: grafana-multi-tenant-proxy
 ---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
   labels:
-    app: loki-multi-tenant-proxy
-  name: loki-multi-tenant-proxy
+    app: grafana-multi-tenant-proxy
+  name: grafana-multi-tenant-proxy
   namespace: grafana
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: loki-multi-tenant-proxy
+      app: grafana-multi-tenant-proxy
   strategy: {}
   template:
     metadata:
       labels:
-        app: loki-multi-tenant-proxy
+        app: grafana-multi-tenant-proxy
     spec:
       containers:
-      - name: loki-multi-tenant-proxy
-        image: angelbarrera92/loki-multi-tenant-proxy:dev
+      - name: grafana-multi-tenant-proxy
+        image: angelbarrera92/grafana-multi-tenant-proxy:dev
         args:
           - "run"
           - "--port=3100"
           - "--loki-server=http://loki.grafana.svc.cluster.local:3100"
-          - "--auth-config=/etc/loki-multi-tenant-proxy/authn.yaml"
+          - "--auth-config=/etc/grafana-multi-tenant-proxy/authn.yaml"
         ports:
           - name: http
             containerPort: 3100
@@ -62,8 +62,8 @@ spec:
         resources: {}
         volumeMounts:
           - name: config
-            mountPath: /etc/loki-multi-tenant-proxy
+            mountPath: /etc/grafana-multi-tenant-proxy
       volumes:
         - name: config
           secret:
-            secretName: loki-multi-tenant-proxy-auth-config
+            secretName: grafana-multi-tenant-proxy-auth-config
